+++ buildSrc/src/main/kotlin/Globals.kt
@@ -1,12 +1,12 @@
 object Versions {
     // Misc and plugins
-    const val GRADLE = "6.8.3"
+    const val GRADLE = "7.0"
     const val KOTLIN = "1.4.32"
     const val BEN_MANES_VERSIONS_PLUGIN = "0.38.0"
     const val DOKKA_PLUGIN = "1.4.30"
 
     // Lib dependencies
-    const val KOTLIN_SERIALIZATION = "1.4.31"
+    const val KOTLIN_SERIALIZATION = KOTLIN
     const val KTOR = "1.5.3"
     const val COROUTINES = "1.4.3"
     const val SLF4J = "1.7.30"